Fixed line continuing bug
Instead of storing the last facing angle and checking if it changed,
we just store a boolean which stores whether or not to continue the
last line segment which is invalidated whenever the `(rotate ...)`
command is used or the push and pop operations are used.

The produced figure now looks as expected.

Fixes #1